# Allow async model moves in model_patcher when --cuda-stream is used and CUDA is available