--- a/ldm_patched/modules/model_patcher.py
+++ b/ldm_patched/modules/model_patcher.py
@@ -22,6 +22,33 @@ PERSISTENT_PATCHES = args.persistent_patches
 if PERSISTENT_PATCHES:
     print("[Experimental] Persistent Patches:", PERSISTENT_PATCHES)
 
+CUDA_STREAM = args.cuda_stream and torch.cuda.is_available()
+
+class AsyncMover:
+    def __init__(self):
+        self.backing_stream = None
+        self.is_streaming = False
+
+    def wait_for_stream(self):
+        if CUDA_STREAM and self.is_streaming and self.backing_stream is not None:
+            torch.cuda.current_stream().wait_stream(self.backing_stream)
+            self.is_streaming = False
+
+    def __call__(self, model, device_to: torch.device):
+        if device_to is None:
+            return
+
+        if CUDA_STREAM:
+            if self.backing_stream is None:
+                self.backing_stream = torch.cuda.Stream()
+            else:
+                self.wait_for_stream()
+
+            with torch.inference_mode(), torch.cuda.stream(self.backing_stream):
+                model.to(device_to, non_blocking=True)
+                self.is_streaming = True
+        else:
+            model.to(device_to)
 
 class PatchStatus:
     def __init__(self):
@@ -63,25 +90,33 @@ class PatchStatus:
 class ModelPatcher:
     def __init__(self, model, load_device, offload_device, size=0, current_device=None, weight_inplace_update=False):
         self.size = size
-        self.model = model
+        self._model = model
         self.patches = {}
         self.backup = {}
         self.object_patches = {}
         self.object_patches_backup = {}
         self.model_options = {"transformer_options": {}}
-        self.model_size()
         self.load_device = load_device
         self.offload_device = offload_device
         self.current_device = self.offload_device if current_device is None else current_device
         self.weight_inplace_update = weight_inplace_update
 
         self.patch_status = PatchStatus()
+        self.async_mover = AsyncMover()
+
+        self.model_size()
+
+    @property
+    def model(self):
+        self.async_mover.wait_for_stream()
+        return self._model
 
     def model_size(self):
         if self.size > 0:
             return self.size
-        model_sd = self.model.state_dict()
-        self.size = ldm_patched.modules.model_management.module_size(self.model)
+        model = self.model
+        model_sd = model.state_dict()
+        self.size = ldm_patched.modules.model_management.module_size(model)
         self.model_keys = set(model_sd.keys())
         return self.size
 
@@ -103,6 +138,7 @@ class ModelPatcher:
         n.model_options = copy.deepcopy(self.model_options)
         n.model_keys = self.model_keys
         n.patch_status = self.patch_status
+        n.async_mover = self.async_mover
 
         return n
 
@@ -243,14 +279,16 @@ class ModelPatcher:
         return sd
 
     def patch_model(self, device_to=None, patch_weights=True):
+        model = self.model
+
         for k in self.object_patches:
-            old = ldm_patched.modules.utils.get_attr(self.model, k)
+            old = ldm_patched.modules.utils.get_attr(model, k)
             if k not in self.object_patches_backup:
                 self.object_patches_backup[k] = old
-            ldm_patched.modules.utils.set_attr_raw(self.model, k, self.object_patches[k])
+            ldm_patched.modules.utils.set_attr_raw(model, k, self.object_patches[k])
 
         if not patch_weights:
-            return self.model
+            return model
 
         if self.patches and self.patch_status.require_patch():
             model_sd = self.model_state_dict()
@@ -272,18 +310,18 @@ class ModelPatcher:
                     temp_weight = weight.to(torch.float32, copy=True)
                 out_weight = self.calculate_weight(self.patches[key], temp_weight, key).to(weight.dtype)
                 if inplace_update:
-                    ldm_patched.modules.utils.copy_to_param(self.model, key, out_weight)
+                    ldm_patched.modules.utils.copy_to_param(model, key, out_weight)
                 else:
-                    ldm_patched.modules.utils.set_attr(self.model, key, out_weight)
+                    ldm_patched.modules.utils.set_attr(model, key, out_weight)
                 del temp_weight
 
             self.patch_status.patch()
 
         if device_to is not None:
-            self.model.to(device_to)
+            self.async_mover(model, device_to)
             self.current_device = device_to
 
-        return self.model
+        return model
 
     def calculate_weight(self, patches, weight, key):
         for p in patches:
@@ -456,26 +494,28 @@ class ModelPatcher:
         return weight
 
     def unpatch_model(self, device_to=None):
+        model = self.model
+
         if self.backup and self.patch_status.require_unpatch():
             keys = list(self.backup.keys())
 
             if self.weight_inplace_update:
                 for k in keys:
-                    ldm_patched.modules.utils.copy_to_param(self.model, k, self.backup[k])
+                    ldm_patched.modules.utils.copy_to_param(model, k, self.backup[k])
             else:
                 for k in keys:
-                    ldm_patched.modules.utils.set_attr(self.model, k, self.backup[k])
+                    ldm_patched.modules.utils.set_attr(model, k, self.backup[k])
 
             self.backup.clear()
             self.patch_status.unpatch()
 
         if device_to is not None:
-            self.model.to(device_to)
+            self.async_mover(model, device_to)
             self.current_device = device_to
 
         keys = list(self.object_patches_backup.keys())
         for k in keys:
-            ldm_patched.modules.utils.set_attr_raw(self.model, k, self.object_patches_backup[k])
+            ldm_patched.modules.utils.set_attr_raw(model, k, self.object_patches_backup[k])
 
         self.object_patches_backup.clear()
 
@@ -483,4 +523,5 @@ class ModelPatcher:
         del self.patches
         del self.object_patches
         del self.model_options
-        self.model = None
+        self.async_mover = None
+        self._model = None