--- a/modules_forge/unet_patcher.py
+++ b/modules_forge/unet_patcher.py
@@ -34,6 +34,7 @@ class UnetPatcher(ModelPatcher):
         n.extra_model_patchers_during_sampling = self.extra_model_patchers_during_sampling.copy()
         n.extra_concat_condition = self.extra_concat_condition
         n.patch_status = self.patch_status
+        n.async_mover = self.async_mover
 
         return n