Allow non-blocking moves of models in model patcher via CUDA streams, improving performance in low VRAM situations, when model ejection and reloading is high.

With this PR, model `to()` calls in `ModelPatcher` are significantly improved in low VRAM situations, and slightly improved under normal circumstances. To test on my 4090 Mobile (16GB VRAM), I ran an image generation of a 32-step 1216x832 image, 15-step upscaling via 4xUltraSharp by 2x, then an ADetailer pass. This requires several model unloads/loads, mainly the UNet, to allow enough VRAM for VAE encoding/decoding. I did several warmup runs before taking numbers, and added some timing code around the model move itself, just to get additional timings:

```
[MODEL PATCHER] Model move took: 0.093s
Moving model(s) has taken 0.12 seconds
```

The `[MODEL PATCHER]` time reflects the duration of the actual model move, while the `Moving model(s)` line captures the total time for the full move operation (which may include patching, reallocation, etc.).

---

### ✅ With Optimization

| Event | `[MODEL PATCHER]` Move Time | `Moving model(s)` Total Time |
|-------|-----------------------------|-------------------------------|
| 1     | 0.005s                      | 0.02s                        |
| 2     | 0.093s                      | 0.12s                        |
| 3     | 0.090s                      | 0.23s                        |
| 4     | 0.089s                      | 0.10s                        |

- **Average `[MODEL PATCHER]` time:** `0.069s`  
- **Average total move time:** `0.1175s`

---

### ❌ Without Optimization

| Event | `[MODEL PATCHER]` Move Time | `Moving model(s)` Total Time |
|-------|-----------------------------|-------------------------------|
| 1     | 0.006s                      | 0.02s                        |
| 2     | 0.712s                      | 0.74s                        |
| 3     | 0.843s                      | 0.87s                        |
| 4     | 0.569s                      | 0.58s                        |

- **Average `[MODEL PATCHER]` time:** `0.5325s`  
- **Average total move time:** `0.5525s`

---

### 🔍 Summary

| Metric                  | Optimized Avg | Unoptimized Avg | Speedup         |
|------------------------|---------------|------------------|-----------------|
| `[MODEL PATCHER]` Time | 0.069s        | 0.5325s          | **~7.7× faster** |
| Total Move Time        | 0.1175s       | 0.5525s          | **~4.7× faster** |

---

Basically, there is a massive reduction in model move times when it does work; the gains are less significant when there's little to no VRAM pressure (as models don't need to be ejected and reloaded).